+++ plugins/frontend/backstage-plugin-github-pull-requests/src/components/PullRequestsStatsCard/PullRequestsStatsCard.test.tsx
@@ -64,7 +64,9 @@ describe('PullRequestsCard', () => {
         </EntityProvider>
       </TestApiProvider>,
     );
-    expect(await screen.findByText('2 months')).toBeInTheDocument();
+    expect(
+      await screen.findByText('1 month, 26 days, 16 hours'),
+    ).toBeInTheDocument();
     expect(await screen.findByText('67%')).toBeInTheDocument();
     expect(await screen.findByText('3309 lines')).toBeInTheDocument();
   });