+++ plugins/frontend/backstage-plugin-github-pull-requests/src/components/usePullRequests.ts
@@ -18,22 +18,22 @@ import { useEffect, useState } from 'react';
 import { useAsyncRetry } from 'react-use';
 import { githubPullRequestsApiRef } from '../api/GithubPullRequestsApi';
 import { useApi, githubAuthApiRef } from '@backstage/core-plugin-api';
-import moment from 'moment';
 import { SearchPullRequestsResponseData } from '../types';
 import { useBaseUrl } from './useBaseUrl';
+import { DateTime } from 'luxon';
 
 export type PullRequest = {
   id: number;
   number: number;
   url: string;
   title: string;
-  updatedTime: string;
-  createdTime: string;
+  updatedTime: string | null;
+  createdTime: string | null;
   state: string;
   draft: boolean;
   merged: string | null;
-  creatorNickname: string;
-  creatorProfileLink: string;
+  creatorNickname?: string | null;
+  creatorProfileLink?: string | null;
   body: string;
 };
 
@@ -54,8 +54,8 @@ export function usePullRequests({
   const [total, setTotal] = useState(0);
   const [page, setPage] = useState(0);
   const [pageSize, setPageSize] = useState(5);
-  const getElapsedTime = (start: string) => {
-    return moment(start).fromNow();
+  const getElapsedTime = (start: string): string | null => {
+    return DateTime.fromISO(start).toRelative();
   };
 
   const {
@@ -108,12 +108,12 @@ export function usePullRequests({
                 id,
                 number,
                 title,
-                body,
+                body: body!,
                 state: pr_state,
                 draft,
                 merged: merged_at,
-                creatorNickname: user.login,
-                creatorProfileLink: user.html_url,
+                creatorNickname: user?.login,
+                creatorProfileLink: user?.html_url,
                 createdTime: getElapsedTime(created_at),
                 updatedTime: getElapsedTime(updated_at),
               }),