+++ utils/roadie-backstage-entity-validator/src/validator.test.js
@@ -1,5 +1,4 @@
 jest.mock('fs');
-
 const validator = require('./validator');
 const { vol } = require('memfs');
 
@@ -129,7 +128,41 @@ spec:
   lifecycle: experimental
   providesApis:
     - sample-service
-    `,
+`,
+      'catalog-info-with-custom-fields.yml': `
+    apiVersion: backstage.io/v1alpha1
+    kind: Component
+    metadata:
+      name: sample-service-5
+      description: |
+        A service for testing Backstage functionality. Configured for GitHub Actions, Sentry, AWS Lambda, Datadog and mis-configured techdocs.
+      annotations:
+        github.com/project-slug: roadiehq/sample-service
+        backstage.io/view-url: https://github.com/RoadieHQ/sample-service/tree/main
+        backstage.io/source-location: url:https://github.com/RoadieHQ/sample-service/tree/main/
+        custom/source-location: custom-field-value:12342
+    spec:
+        type: service
+        owner: user:dtuite
+        lifecycle: experimental
+`,
+      'catalog-info-with-custom-fields-with-errors.yml': `
+        apiVersion: backstage.io/v1alpha1
+        kind: Component
+        metadata:
+          name: sample-service-5
+          description: |
+            A service for testing Backstage functionality. Configured for GitHub Actions, Sentry, AWS Lambda, Datadog and mis-configured techdocs.
+          annotations:
+            github.com/project-slug: roadiehq/sample-service
+            backstage.io/view-url: https://github.com/RoadieHQ/sample-service/tree/main
+            backstage.io/source-location: url:https://github.com/RoadieHQ/sample-service/tree/main/
+            custom/source-location: custom-field-value:123error
+        spec:
+            type: service
+            owner: user:dtuite
+            lifecycle: experimental
+`,
       'template-v2-entity.yml': `
 apiVersion: backstage.io/v1beta2
 kind: Template
@@ -160,6 +193,35 @@ spec:
       path: foo.txt
       outputPath: foo.zip
 `,
+      'custom-validation-schema.json': `
+{
+  "$schema": "http://json-schema.org/draft-07/schema",
+  "$id": "Custom Entity metadata annotations",
+  "description": "Individual annotation format validations",
+  "type": "object",
+  "required": ["metadata"],
+  "additionalProperties": true,
+  "properties": {
+    "metadata": {
+      "type": "object",
+      "properties": {
+        "annotations": {
+          "type": "object",
+          "allOf": [
+            {
+              "properties": {
+                "custom/source-location": {
+                  "type": "string",
+                  "pattern": "^custom-field-value:\\\\d*$"
+                }
+              }
+            }
+          ]
+        }
+      }
+    }
+  }
+}`,
     });
   });
   afterEach(() => {
@@ -266,6 +328,30 @@ spec:
       await expect(
         validator.validateFromFile('template-v3-entity.yml'),
       ).resolves.toBeUndefined();
+    });
+  });
+
+  describe('customValidationSchema', () => {
+    it('should successfully validate against custom annotation schema', async () => {
+      await expect(
+        validator.validateFromFile(
+          'catalog-info-with-custom-fields.yml',
+          false,
+          'custom-validation-schema.json',
+        ),
+      ).resolves.toBeUndefined();
+    });
+
+    it('should throw validate error when validating against custom annotation schema', async () => {
+      await expect(
+        validator.validateFromFile(
+          'catalog-info-with-custom-fields-with-errors.yml',
+          false,
+          'custom-validation-schema.json',
+        ),
+      ).rejects.toThrowError(
+        'Error: Malformed annotation, /metadata/annotations/custom~1source-location must match pattern "^custom-field-value:\\d*$"',
+      );
     });
   });
 });