+++ plugins/frontend/backstage-plugin-jira/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @roadiehq/backstage-plugin-jira
 
+## 2.4.4
+
+### Patch Changes
+
+- 0fe673e3: Version bumps
+
+## 2.4.3
+
+### Patch Changes
+
+- 85620abc: Bump packages to backstage version 1.13.0
+
 ## 2.4.2
 
 ### Patch Changes