+++ plugins/frontend/backstage-plugin-prometheus/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-prometheus
 
+## 2.4.6
+
+### Patch Changes
+
+- 85620abc: Bump packages to backstage version 1.13.0
+
 ## 2.4.5
 
 ### Patch Changes